workflow/requirements.txt
flytekitplugins-bacalhau
bacalhau-sdk
boto3
python-dotenv